replacing vector by array for wrong guesses
The maximum wrong guesses allowed are 6; if we use vector, we have a
data structure allocated on the heap and that allows n values. Using
array, we can store it in the stack frame and have only 6 positions.

The value (6) is actually stored in a constant (MAX_TRIES), so if more
or less tries are allowed one needs only to adjust this const.

+ const parameters in some functions.